Rename examples to example.
This keeps the directory inline with our guidelines on singular
names for directories.
