+++ example/servicegraph/BUILD
@@ -1,4 +1,4 @@
-package(default_visibility = ["//examples/servicegraph:__subpackages__"])
+package(default_visibility = ["//example/servicegraph:__subpackages__"])
 
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
