+++ example/servicegraph/cmd/demosvc/BUILD
