+++ example/servicegraph/dot/BUILD
@@ -1,4 +1,4 @@
-package(default_visibility = ["//examples/servicegraph/cmd:__subpackages__"])
+package(default_visibility = ["//example/servicegraph/cmd:__subpackages__"])
 
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
@@ -8,6 +8,6 @@ go_library(
         "dot.go",
     ],
     deps = [
-        "//examples/servicegraph:go_default_library",
+        "//example/servicegraph:go_default_library",
     ],
 )
