+++ example/servicegraph/promgen/BUILD
@@ -1,4 +1,4 @@
-package(default_visibility = ["//examples/servicegraph/cmd:__subpackages__"])
+package(default_visibility = ["//example/servicegraph/cmd:__subpackages__"])
 
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
@@ -8,7 +8,7 @@ go_library(
         "promgen.go",
     ],
     deps = [
-        "//examples/servicegraph:go_default_library",
+        "//example/servicegraph:go_default_library",
         "@com_github_prometheus_client_golang//api/prometheus:go_default_library",
         "@com_github_prometheus_common//model:go_default_library",
     ],
