--- examples/servicegraph/cmd/server/BUILD
@@ -1,21 +1,0 @@
-package(default_visibility = ["//examples/servicegraph:__subpackages__"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_binary")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "main.go",
-    ],
-    deps = [
-        "//examples/servicegraph:go_default_library",
-        "//examples/servicegraph/dot:go_default_library",
-        "//examples/servicegraph/promgen:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "servicegraph",
-    data = ["//examples/servicegraph/js:viz/viz-lite.js"],
-    library = ":go_default_library",
-)
